added wumpus.py
provided wumpus envronment of the cave system/caves. Each cave can have a pit, the wumpus, or the gold. The str built-in function is overloaded so that Tiles and and the Grid can be printed properly. The printed Grid is an NxM of boolean vectors. In each vector, there are 3 boolean values. The leftmost indicates pit presence, the middle wumpus presence, and the rightmost gold presence.